backend/kater/requirements.txt
django==5.0.6
django-environ==0.9.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2
PyJWT